+++ extensions/spotify-player/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Spotify Player Changelog
 
+## [Fix Search Feature] - 2025-07-02
+
+- Fixed an issue with the search functionality.
+
 ## [Add Option to View the Lyrics of the Song playing] - 2025-06-30
 
 - Add new command 'Find Lyrics'.